+++ src/lib.rs
@@ -4,4 +4,6 @@ pub mod map;
 pub mod math;
 pub mod object;
 pub mod offline_map;
+pub mod room_coordinate;
+pub mod room_xy;
 pub mod sparse_cost_matrix;